refactor: move cli command handlers into runtime module
