+++ findings.md
@@ -8,13 +8,15 @@
 - Session name/path expansion was a good first reusable extraction target because it had clear boundaries and minimal side effects.
 - Command alias matching/dispatch is another clean extraction point that reduces CLI frontend coupling.
 - Option parsing is a third good extraction target because it is self-contained and shared by multiple command paths.
+- Moving command handlers into a runtime module gives clearer separation between CLI orchestration (`atch.c`) and command behavior implementation.
 
 ## Decisions
 - Keep core behavior unchanged; avoid moving large code blocks in first refactor.
-- Preserve current CLI parser/dispatcher in `atch.c` while progressively extracting reusable modules.
-- Treat this as phase 1+2 internal modularization toward future library split.
+- Preserve legacy CLI semantics while progressively extracting reusable modules.
+- Treat this as phase 1+2 internal modularization toward future library split and richer frontends.
 
 ## Notes
 - `atch_expand_session_name_dup` centralizes bare-name expansion and cache-dir creation behavior for future frontends.
 - `atch_resolve_command` centralizes alias policy (`list/l/ls`, `attach/a`, etc.) for reuse outside the CLI binary.
-- `atch_parse_options` now centralizes option semantics and validation, reducing direct coupling in `atch.c`.
+- `atch_parse_options` centralizes option semantics and validation, reducing direct coupling in `atch.c`.
+- `atch_cli_runtime` now encapsulates command handlers and shared runtime helpers for potential alternate CLIs/managers.
